+++ tests/Unit/Field/FieldTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace SmartAssert\ServiceRequest\Tests\Unit\Field;
 
 use PHPUnit\Framework\TestCase;
-use SmartAssert\ServiceRequest\Field\FieldInterface;
+use SmartAssert\ServiceRequest\Parameter\FieldInterface;
 use SmartAssert\ServiceRequest\Tests\DataProvider\FieldDataProviderTrait;
 
 class FieldTest extends TestCase
